Split embed methods for Vertex/AI Studio
